renamed test file, got spark to work with iceberg table
